pyproject.toml
[tool.poetry]
name = "warehouse_api"
version = "0.1.0"
description = ""
authors = ["VanderSVan <101255612+VanderSVan@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.10"
SQLAlchemy = "^1.4.41"
alembic = "^1.8.1"
fastapi = "^0.84.0"
pydantic = {extras = ["email", "dotenv"], version = "^1.10.2"}
asyncpg = "^0.26.0"
uvicorn = {extras = ["standard"], version = "^0.18.3"}
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = "^1.7.4"
blake2signer = "^2.5.0"
bcrypt = "^4.0.0"
fastapi-mail = "^1.1.4"
celery = "^5.2.7"
redis = "^4.3.4"
flower = "^1.2.0"
python-multipart = "^0.0.5"
psycopg2-binary = "^2.9.3"
pytest = "^7.1.3"
requests = "^2.28.1"
aioredis = "^2.0.1"
httpx = "^0.23.0"
pytest-asyncio = "^0.19.0"
orjson = "^3.8.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
